Move form components to a form folder.
